requirements.txt
openai==0.28.0
pandas==2.0.3
plotly==5.24.1
python_dateutil==2.9.0
retrying==1.3.4
streamlit==1.40.0
tqdm==4.67.0
